+++ src/main/java/duke/Storage.java
@@ -1,3 +1,9 @@
+package duke;
+
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.Parser;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
