+++ src/main/java/duke/Ui.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.task.Task;
+import duke.task.TaskList;
 import java.util.Scanner;
 
 public class Ui {
