+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,17 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.Ui;
+import duke.command.Command;
+
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.task.Event;
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.task.Todo;
+
 import java.io.IOException;
 import java.util.ArrayList;
