+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
 import java.util.ArrayList;
 
 public class ByeCommand extends Command {
